requirements.txt
# requirements.txt
streamlit>=1.28.0
pillow
pandas
matplotlib
python-dotenv
requests
google-generativeai
fuzzywuzzy[speedup]
plotly
